backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
PyJWT==2.8.1
bcrypt==4.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
